Create ReadMe
Creative Lab
Background
Due to the global warming and other climate crises, the temperature of the air will increase and decrease rapidly in future. Human need to open the shell to protect crop and go into the refuge to avoid the temperature disaster when the temperature starts to change. Therefore, human create an alarm to warn all people.

Working Principle
“ According to the research of the temperature disaster from scientists, the air temperature will steadily increate/decrease 3 celsius per minute. ”

According to the principle of the temperature change, whether the temperature change exceeds 0.4 Celsius every 10 seconds becomes the reference standard. If the alarm detects four time temperature changes more than standard change in one minute, it will turn to warning situation. The alarm has three LEDs to indicate different situations:
Green LED: when the changes larger than the reference standard are less than 4 times. It means that the temperature is changing in a safe range. People can enjoy their daily life.
Yellow LED: when the changes larger than the reference standard are more than 4 times, but less than 6 times. It means the temperature disaster may come soon, and people need to pay attention to the dangerous situation.
Red LED: when the changes larger than the reference standard are more than 6 times. It means that the temperature disaster is coming. People need to take protective measures.